refactor(service): 拆分经验治理留痕辅助逻辑
